requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
# duckdb
pandas
openai
langchain-groq
# duckduckgo-search
mysql-connector-python
SQLAlchemy
validator
unstructured
pytube
numexpr
streamlit 
# langchain_faiss
validators==0.28.1
streamlit
google-generativeai
python-dotenv
langchain
PyPDF2
chromadb
faiss-cpu
pdf2image
# sqlite3
google-generativeai